Refatorado SetV1 Node
